Fix root of c64 coordinate system
Back when we imlemented this, we guessed, that the X-root of the coordinate system is probably
exactly in the middle of the left border, apparently this was not correct, as sprites were displayed
two pixels too far left.

So we shift the root of the coordinate system a bit to the left, and it matches the sprite
coordinates again.